+++ src/TrashLib/Repo/RepoPaths.cs
@@ -4,5 +4,7 @@ namespace TrashLib.Repo;
 
 public record RepoPaths(
     IReadOnlyCollection<IDirectoryInfo> RadarrCustomFormatPaths,
-    IReadOnlyCollection<IDirectoryInfo> SonarrReleaseProfilePaths
+    IReadOnlyCollection<IDirectoryInfo> SonarrReleaseProfilePaths,
+    IReadOnlyCollection<IDirectoryInfo> RadarrQualityPaths,
+    IReadOnlyCollection<IDirectoryInfo> SonarrQualityPaths
 ) : IRepoPaths;