+++ src/TrashLib/Services/Radarr/QualityDefinition/IRadarrQualityDefinitionGuideParser.cs
@@ -1,7 +1,0 @@
-namespace TrashLib.Services.Radarr.QualityDefinition;
-
-public interface IRadarrQualityDefinitionGuideParser
-{
-    Task<string> GetMarkdownData();
-    IDictionary<RadarrQualityDefinitionType, List<RadarrQualityData>> ParseMarkdown(string markdown);
-}