+++ src/TrashLib/Services/Radarr/QualityDefinition/IRadarrQualityGuideParser.cs
@@ -1,0 +1,6 @@
+namespace TrashLib.Services.Radarr.QualityDefinition;
+
+public interface IRadarrQualityGuideParser
+{
+    ICollection<RadarrQualityData> GetQualities();
+}