+++ src/TrashLib/Services/Radarr/QualityDefinition/RadarrQualityDefinitionType.cs
@@ -1,6 +1,0 @@
-namespace TrashLib.Services.Radarr.QualityDefinition;
-
-public enum RadarrQualityDefinitionType
-{
-    Movie
-}