+++ src/TrashLib/Services/Sonarr/QualityDefinition/ISonarrQualityGuideParser.cs
@@ -1,0 +1,6 @@
+namespace TrashLib.Services.Sonarr.QualityDefinition;
+
+public interface ISonarrQualityGuideParser
+{
+    ICollection<SonarrQualityData> GetQualities();
+}