+++ src/TrashLib/Services/Sonarr/QualityDefinition/SonarrQualityDefinitionType.cs
@@ -1,8 +1,0 @@
-namespace TrashLib.Services.Sonarr.QualityDefinition;
-
-public enum SonarrQualityDefinitionType
-{
-    Anime,
-    Series,
-    Hybrid
-}