+++ src/TrashLib/Services/Sonarr/QualityDefinition/SonarrQualityDefinitionUpdater.cs
@@ -1,6 +1,8 @@
 using System.Text.RegularExpressions;
 using CliFx.Infrastructure;
+using Common.Extensions;
 using Serilog;
+using TrashLib.Services.Common.QualityDefinition;
 using TrashLib.Services.Sonarr.Api;
 using TrashLib.Services.Sonarr.Api.Objects;
 using TrashLib.Services.Sonarr.Config;
@@ -9,91 +11,117 @@ namespace TrashLib.Services.Sonarr.QualityDefinition;
 
 internal class SonarrQualityDefinitionUpdater : ISonarrQualityDefinitionUpdater
 {
+    private readonly ILogger _log;
     private readonly ISonarrApi _api;
     private readonly IConsole _console;
-    private readonly ISonarrQualityDefinitionGuideParser _parser;
+    private readonly ISonarrQualityGuideParser _parser;
     private readonly Regex _regexHybrid = new(@"720|1080", RegexOptions.Compiled);
 
     public SonarrQualityDefinitionUpdater(
         ILogger logger,
-        ISonarrQualityDefinitionGuideParser parser,
+        ISonarrQualityGuideParser parser,
         ISonarrApi api,
         IConsole console)
     {
-        Log = logger;
+        _log = logger;
         _parser = parser;
         _api = api;
         _console = console;
     }
 
-    private ILogger Log { get; }
+    private SonarrQualityData? GetQualityOrError(ICollection<SonarrQualityData> qualityDefinitions, string type)
+    {
+        var quality = qualityDefinitions.FirstOrDefault(x => x.Type.EqualsIgnoreCase(type));
+        if (quality is null)
+        {
+            _log.Error(
+                "The following quality definition is required for hybrid, but was not found in the guide: {Type}",
+                type);
+        }
+
+        return quality;
+    }
 
     public async Task Process(bool isPreview, SonarrConfiguration config)
     {
-        Log.Information("Processing Quality Definition: {QualityDefinition}", config.QualityDefinition);
-        var qualityDefinitions = _parser.ParseMarkdown(await _parser.GetMarkdownData());
-        List<SonarrQualityData> selectedQuality;
+        _log.Information("Processing Quality Definition: {QualityDefinition}", config.QualityDefinition);
+        var qualityDefinitions = _parser.GetQualities();
+        var qualityTypeInConfig = config.QualityDefinition;
 
-        if (config.QualityDefinition == SonarrQualityDefinitionType.Hybrid)
+        // var qualityDefinitions = _parser.ParseMarkdown(await _parser.GetMarkdownData());
+        SonarrQualityData? selectedQuality;
+
+        if (config.QualityDefinition.EqualsIgnoreCase("hybrid"))
         {
-            selectedQuality = BuildHybridQuality(qualityDefinitions[SonarrQualityDefinitionType.Anime],
-                qualityDefinitions[SonarrQualityDefinitionType.Series]);
+            var animeQuality = GetQualityOrError(qualityDefinitions, "anime");
+            var seriesQuality = GetQualityOrError(qualityDefinitions, "series");
+            if (animeQuality is null || seriesQuality is null)
+            {
+                return;
+            }
+
+            selectedQuality = BuildHybridQuality(animeQuality.Qualities, seriesQuality.Qualities);
         }
         else
         {
-            selectedQuality = qualityDefinitions[config.QualityDefinition!.Value];
+            selectedQuality = qualityDefinitions
+                .FirstOrDefault(x => x.Type.EqualsIgnoreCase(qualityTypeInConfig));
+
+            if (selectedQuality == null)
+            {
+                _log.Error("The specified quality definition type does not exist: {Type}", qualityTypeInConfig);
+                return;
+            }
         }
 
         if (isPreview)
         {
-            PrintQualityPreview(selectedQuality);
+            PrintQualityPreview(selectedQuality.Qualities);
             return;
         }
 
-        await ProcessQualityDefinition(selectedQuality);
+        await ProcessQualityDefinition(selectedQuality.Qualities);
     }
 
-    private List<SonarrQualityData> BuildHybridQuality(IReadOnlyCollection<SonarrQualityData> anime,
-        IEnumerable<SonarrQualityData> series)
+    private SonarrQualityData BuildHybridQuality(
+        IReadOnlyCollection<QualityItem> anime,
+        IReadOnlyCollection<QualityItem> series)
     {
         // todo Verify anime & series are the same length? Probably not, because we might not care about some rows anyway.
-        Log.Information(
+        _log.Information(
             "Notice: Hybrid only functions on 720/1080 qualities and uses non-anime values for the rest (e.g. 2160)");
 
-        var hybrid = new List<SonarrQualityData>();
+        var hybrid = new List<QualityItem>();
         foreach (var left in series)
         {
             // Any qualities that anime doesn't care about get immediately added from Series quality
-            var match = _regexHybrid.Match(left.Name);
+            var match = _regexHybrid.Match(left.Quality);
             if (!match.Success)
             {
-                Log.Debug("Using 'Series' Quality For: {QualityName}", left.Name);
+                _log.Debug("Using 'Series' Quality For: {QualityName}", left.Quality);
                 hybrid.Add(left);
                 continue;
             }
 
             // If there's a quality in Series that Anime doesn't know about, we add the Series quality
-            var right = anime.FirstOrDefault(row => row.Name == left.Name);
+            var right = anime.FirstOrDefault(row => row.Quality == left.Quality);
             if (right == null)
             {
-                Log.Error("Could not find matching anime quality for series quality named {QualityName}",
-                    left.Name);
+                _log.Error("Could not find matching anime quality for series quality named {QualityName}",
+                    left.Quality);
                 hybrid.Add(left);
                 continue;
             }
 
-            hybrid.Add(new SonarrQualityData
-            {
-                Name = left.Name,
-                Min = Math.Min(left.Min, right.Min),
-                Max = Math.Max(left.Max, right.Max)
-            });
+            hybrid.Add(new QualityItem(left.Quality,
+                Math.Min(left.Min, right.Min),
+                Math.Max(left.Max, right.Max)));
         }
 
-        return hybrid;
+        return new SonarrQualityData("", "hybrid", hybrid);
     }
 
-    private void PrintQualityPreview(IEnumerable<SonarrQualityData> quality)
+    private void PrintQualityPreview(IEnumerable<QualityItem> quality)
     {
         _console.Output.WriteLine("");
         const string format = "{0,-20} {1,-10} {2,-15}";
@@ -102,22 +130,22 @@ internal class SonarrQualityDefinitionUpdater : ISonarrQualityDefinitionUpdater
 
         foreach (var q in quality)
         {
-            _console.Output.WriteLine(format, q.Name, q.AnnotatedMin, q.AnnotatedMax);
+            _console.Output.WriteLine(format, q.Quality, q.AnnotatedMin, q.AnnotatedMax);
         }
 
         _console.Output.WriteLine("");
     }
 
-    private async Task ProcessQualityDefinition(IEnumerable<SonarrQualityData> guideQuality)
+    private async Task ProcessQualityDefinition(IEnumerable<QualityItem> guideQuality)
     {
         var serverQuality = await _api.GetQualityDefinition();
         await UpdateQualityDefinition(serverQuality, guideQuality);
     }
 
     private async Task UpdateQualityDefinition(IReadOnlyCollection<SonarrQualityDefinitionItem> serverQuality,
-        IEnumerable<SonarrQualityData> guideQuality)
+        IEnumerable<QualityItem> guideQuality)
     {
-        static bool QualityIsDifferent(SonarrQualityDefinitionItem a, SonarrQualityData b)
+        static bool QualityIsDifferent(SonarrQualityDefinitionItem a, QualityItem b)
         {
             return b.IsMinDifferent(a.MinSize) ||
                    b.IsMaxDifferent(a.MaxSize);
@@ -126,10 +154,10 @@ internal class SonarrQualityDefinitionUpdater : ISonarrQualityDefinitionUpdater
         var newQuality = new List<SonarrQualityDefinitionItem>();
         foreach (var qualityData in guideQuality)
         {
-            var entry = serverQuality.FirstOrDefault(q => q.Quality?.Name == qualityData.Name);
+            var entry = serverQuality.FirstOrDefault(q => q.Quality?.Name == qualityData.Quality);
             if (entry == null)
             {
-                Log.Warning("Server lacks quality definition for {Quality}; it will be skipped", qualityData.Name);
+                _log.Warning("Server lacks quality definition for {Quality}; it will be skipped", qualityData.Quality);
                 continue;
             }
 
@@ -143,12 +171,12 @@ internal class SonarrQualityDefinitionUpdater : ISonarrQualityDefinitionUpdater
             entry.MaxSize = qualityData.MaxForApi;
             newQuality.Add(entry);
 
-            Log.Debug("Setting Quality " +
-                      "[Name: {Name}] [Source: {Source}] [Min: {Min}] [Max: {Max}]",
+            _log.Debug("Setting Quality " +
+                       "[Name: {Name}] [Source: {Source}] [Min: {Min}] [Max: {Max}]",
                 entry.Quality?.Name, entry.Quality?.Source, entry.MinSize, entry.MaxSize);
         }
 
         await _api.UpdateQualityDefinition(newQuality);
-        Log.Information("Number of updated qualities: {Count}", newQuality.Count);
+        _log.Information("Number of updated qualities: {Count}", newQuality.Count);
     }
 }